Mix !might and !brill into a gross slurry
I have no idea if this is going to be merged but I did the work for
it anyway. (It might be a bit cleaner if the str/int bonuses were
removed; you could do this by either adding a DUR_POTENCY or by
moving the str bonus from DUR_MIGHT into a secondary secret duration.
I like the latter but I dont' feel strongly.)